OpenRISC port step 13: Function result as argument
If a function is called within an argument list of another function:

    func_outer(1, func_inner(5, 6), 3);

and the arguments are passed in registers, the inner arguments overwrite
the outer arguments because the same registers are used. There is no
problem if the arguments are passed on a stack in memory.

To solve the problem, implement `emit_pre_call()` to save the registers
with the outer arguments and restore them in a modified `emit_call()`.